Separate out the WindowInfo and ApplicationInfo mutators so we don't play silly tricks with the linker and don't expose the entry points.
