+++ src/miral/window_info.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2015-2016 Canonical Ltd.
+ * Copyright © 2015-2020 Canonical Ltd.
  *
  * This program is free software: you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 or 3 as
@@ -18,124 +18,13 @@
 
 #include "miral/window_info.h"
 #include "window_info_defaults.h"
+#include "window_info_internal.h"
 
 #include <mir/scene/surface.h>
 
 #include <limits>
 
 using namespace mir::geometry;
-
-namespace
-{
-template<typename Value>
-auto optional_value_or_default(mir::optional_value<Value> const& optional_value, Value default_ = Value{})
--> Value
-{
-    return optional_value.is_set() ? optional_value.value() : default_;
-}
-
-unsigned clamp_dim(unsigned dim)
-{
-    return std::min<unsigned long>(std::numeric_limits<long>::max(), dim);
-}
-
-// For our convenience when doing calculations we clamp the dimensions of the aspect ratio
-// so they will fit into a long without overflow.
-miral::WindowInfo::AspectRatio clamp(miral::WindowInfo::AspectRatio const& source)
-{
-    return {clamp_dim(source.width), clamp_dim(source.height)};
-}
-
-miral::Width clamp(miral::Width const& source)
-{
-    return std::min(miral::default_max_width, std::max(miral::default_min_width, source));
-}
-
-miral::Height clamp(miral::Height const& source)
-{
-    return std::min(miral::default_max_height, std::max(miral::default_min_height, source));
-}
-}
-
-miral::Width  const miral::default_min_width{0};
-miral::Height const miral::default_min_height{0};
-miral::Width  const miral::default_max_width{std::numeric_limits<int>::max()};
-miral::Height const miral::default_max_height{std::numeric_limits<int>::max()};
-miral::DeltaX const miral::default_width_inc{1};
-miral::DeltaY const miral::default_height_inc{1};
-miral::WindowInfo::AspectRatio const miral::default_min_aspect_ratio{
-    clamp(WindowInfo::AspectRatio{0U, std::numeric_limits<unsigned>::max()})};
-miral::WindowInfo::AspectRatio const miral::default_max_aspect_ratio{
-    clamp(WindowInfo::AspectRatio{std::numeric_limits<unsigned>::max(), 0U})};
-
-struct miral::WindowInfo::Self
-{
-    Self(Window window, WindowSpecification const& params);
-    Self();
-
-    Window window;
-    std::string name;
-    MirWindowType type;
-    MirWindowState state;
-    mir::geometry::Rectangle restore_rect;
-    Window parent;
-    std::vector <Window> children;
-    mir::geometry::Width min_width;
-    mir::geometry::Height min_height;
-    mir::geometry::Width max_width;
-    mir::geometry::Height max_height;
-    MirOrientationMode preferred_orientation;
-    MirPointerConfinementState confine_pointer;
-
-    mir::geometry::DeltaX width_inc;
-    mir::geometry::DeltaY height_inc;
-    AspectRatio min_aspect;
-    AspectRatio max_aspect;
-    mir::optional_value<int> output_id;
-    MirShellChrome shell_chrome;
-    MirDepthLayer depth_layer;
-    MirPlacementGravity attached_edges;
-    mir::optional_value<mir::geometry::Rectangle> exclusive_rect;
-    std::shared_ptr<void> userdata;
-};
-
-miral::WindowInfo::Self::Self(Window window, WindowSpecification const& params) :
-    window{window},
-    name{params.name().value()},
-    type{optional_value_or_default(params.type(), mir_window_type_normal)},
-    state{optional_value_or_default(params.state(), mir_window_state_restored)},
-    restore_rect{params.top_left().value(), params.size().value()},
-    min_width{optional_value_or_default(params.min_width(), miral::default_min_width)},
-    min_height{optional_value_or_default(params.min_height(), miral::default_min_height)},
-    max_width{optional_value_or_default(params.max_width(), miral::default_max_width)},
-    max_height{optional_value_or_default(params.max_height(), miral::default_max_height)},
-    preferred_orientation{optional_value_or_default(params.preferred_orientation(), mir_orientation_mode_any)},
-    confine_pointer(optional_value_or_default(params.confine_pointer(), mir_pointer_unconfined)),
-    width_inc{optional_value_or_default(params.width_inc(), default_width_inc)},
-    height_inc{optional_value_or_default(params.height_inc(), default_height_inc)},
-    min_aspect(optional_value_or_default(params.min_aspect(), default_min_aspect_ratio)),
-    max_aspect(optional_value_or_default(params.max_aspect(), default_max_aspect_ratio)),
-    shell_chrome(optional_value_or_default(params.shell_chrome(), mir_shell_chrome_normal)),
-    depth_layer(optional_value_or_default(params.depth_layer(), mir_depth_layer_application)),
-    attached_edges(optional_value_or_default(params.attached_edges(), mir_placement_gravity_center))
-{
-    if (params.output_id().is_set())
-        output_id = params.output_id().value();
-
-    if (params.exclusive_rect().is_set())
-        exclusive_rect = params.exclusive_rect().value();
-
-    if (params.userdata().is_set())
-        userdata = params.userdata().value();
-}
-
-miral::WindowInfo::Self::Self() :
-    type{mir_window_type_normal},
-    state{mir_window_state_unknown},
-    preferred_orientation{mir_orientation_mode_any},
-    shell_chrome{mir_shell_chrome_normal}
-{
-}
 
 miral::WindowInfo::WindowInfo() :
     self{std::make_unique<Self>()}
@@ -416,19 +305,9 @@ auto miral::WindowInfo::type() const -> MirWindowType
     return self->type;
 }
 
-void miral::WindowInfo::type(MirWindowType type)
-{
-    self->type = type;
-}
-
 auto miral::WindowInfo::state() const -> MirWindowState
 {
     return self->state;
-}
-
-void miral::WindowInfo::state(MirWindowState state)
-{
-    self->state = state;
 }
 
 auto miral::WindowInfo::restore_rect() const -> mir::geometry::Rectangle
@@ -436,19 +315,9 @@ auto miral::WindowInfo::restore_rect() const -> mir::geometry::Rectangle
     return self->restore_rect;
 }
 
-void miral::WindowInfo::restore_rect(mir::geometry::Rectangle const& restore_rect)
-{
-    self->restore_rect = restore_rect;
-}
-
 auto miral::WindowInfo::parent() const -> Window
 {
     return self->parent;
-}
-
-void miral::WindowInfo::parent(Window const& parent)
-{
-    self->parent = parent;
 }
 
 auto miral::WindowInfo::children() const -> std::vector <Window> const&
@@ -456,33 +325,9 @@ auto miral::WindowInfo::children() const -> std::vector <Window> const&
     return self->children;
 }
 
-void miral::WindowInfo::add_child(Window const& child)
-{
-    self->children.push_back(child);
-}
-
-void miral::WindowInfo::remove_child(Window const& child)
-{
-    auto& siblings = self->children;
-
-    for (auto i = begin(siblings); i != end(siblings); ++i)
-    {
-        if (child == *i)
-        {
-            siblings.erase(i);
-            break;
-        }
-    }
-}
-
 auto miral::WindowInfo::min_width() const -> mir::geometry::Width
 {
     return self->min_width;
-}
-
-void miral::WindowInfo::min_width(mir::geometry::Width min_width)
-{
-    self->min_width = clamp(min_width);
 }
 
 auto miral::WindowInfo::min_height() const -> mir::geometry::Height
@@ -490,29 +335,14 @@ auto miral::WindowInfo::min_height() const -> mir::geometry::Height
     return self->min_height;
 }
 
-void miral::WindowInfo::min_height(mir::geometry::Height min_height)
-{
-    self->min_height = clamp(min_height);
-}
-
 auto miral::WindowInfo::max_width() const -> mir::geometry::Width
 {
     return self->max_width;
 }
 
-void miral::WindowInfo::max_width(mir::geometry::Width max_width)
-{
-    self->max_width = clamp(max_width);
-}
-
 auto miral::WindowInfo::max_height() const -> mir::geometry::Height
 {
     return self->max_height;
-}
-
-void miral::WindowInfo::max_height(mir::geometry::Height max_height)
-{
-    self->max_height = clamp(max_height);
 }
 
 auto miral::WindowInfo::userdata() const -> std::shared_ptr<void>
@@ -530,19 +360,9 @@ auto miral::WindowInfo::width_inc() const -> mir::geometry::DeltaX
     return self->width_inc;
 }
 
-void miral::WindowInfo::width_inc(DeltaX width_inc)
-{
-    self->width_inc = width_inc;
-}
-
 auto miral::WindowInfo::height_inc() const -> mir::geometry::DeltaY
 {
     return self->height_inc;
-}
-
-void miral::WindowInfo::height_inc(mir::geometry::DeltaY height_inc)
-{
-    self->height_inc = height_inc;
 }
 
 auto miral::WindowInfo::min_aspect() const -> AspectRatio
@@ -550,19 +370,9 @@ auto miral::WindowInfo::min_aspect() const -> AspectRatio
     return self->min_aspect;
 }
 
-void miral::WindowInfo::min_aspect(AspectRatio min_aspect)
-{
-    self->min_aspect = clamp(min_aspect);
-}
-
 auto miral::WindowInfo::max_aspect() const -> AspectRatio
 {
     return self->max_aspect;
-}
-
-void miral::WindowInfo::max_aspect(AspectRatio max_aspect)
-{
-    self->max_aspect = clamp(max_aspect);
 }
 
 bool miral::WindowInfo::has_output_id() const
@@ -575,11 +385,6 @@ auto miral::WindowInfo::output_id() const -> int
     return self->output_id.value();
 }
 
-void miral::WindowInfo::output_id(mir::optional_value<int> output_id)
-{
-    self->output_id = output_id;
-}
-
 auto miral::WindowInfo::window() const -> Window&
 {
     return self->window;
@@ -590,19 +395,9 @@ auto miral::WindowInfo::preferred_orientation() const -> MirOrientationMode
     return self->preferred_orientation;
 }
 
-void miral::WindowInfo::preferred_orientation(MirOrientationMode preferred_orientation)
-{
-    self->preferred_orientation = preferred_orientation;
-}
-
 auto miral::WindowInfo::confine_pointer() const -> MirPointerConfinementState
 {
     return self->confine_pointer;
-}
-
-void miral::WindowInfo::confine_pointer(MirPointerConfinementState confinement)
-{
-    self->confine_pointer = confinement;
 }
 
 auto miral::WindowInfo::shell_chrome() const -> MirShellChrome
@@ -610,19 +405,9 @@ auto miral::WindowInfo::shell_chrome() const -> MirShellChrome
     return self->shell_chrome;
 }
 
-void miral::WindowInfo::shell_chrome(MirShellChrome chrome)
-{
-    self->shell_chrome = chrome;
-}
-
 auto miral::WindowInfo::name() const -> std::string
 {
     return self->name;
-}
-
-void miral::WindowInfo::name(std::string const& name)
-{
-    self->name = name;
 }
 
 auto miral::WindowInfo::depth_layer() const -> MirDepthLayer
@@ -630,29 +415,14 @@ auto miral::WindowInfo::depth_layer() const -> MirDepthLayer
     return self->depth_layer;
 }
 
-void miral::WindowInfo::depth_layer(MirDepthLayer depth_layer)
-{
-    self->depth_layer = depth_layer;
-}
-
 auto miral::WindowInfo::attached_edges() const -> MirPlacementGravity
 {
     return self->attached_edges;
 }
 
-void miral::WindowInfo::attached_edges(MirPlacementGravity edges)
-{
-    self->attached_edges = edges;
-}
-
 auto miral::WindowInfo::exclusive_rect() const -> mir::optional_value<mir::geometry::Rectangle>
 {
     return self->exclusive_rect;
-}
-
-void miral::WindowInfo::exclusive_rect(mir::optional_value<mir::geometry::Rectangle> const& rect)
-{
-    self->exclusive_rect = rect;
 }
 
 auto miral::WindowInfo::clip_area() const -> mir::optional_value<mir::geometry::Rectangle>
@@ -682,11 +452,4 @@ auto miral::WindowInfo::application_id() const -> std::string
         return surface->application_id();
     else
         return "";
-}
-
-void miral::WindowInfo::application_id(std::string const& application_id)
-{
-    std::shared_ptr<mir::scene::Surface> surface = window();
-    if (surface)
-        return surface->set_application_id(application_id);
 }
